+++ packages/forms/CHANGELOG.md
@@ -1,5 +1,48 @@
 # @px-ui/forms
 
+## 11.0.2
+
+### Patch Changes
+
+- Updated dependencies [ab2d61c]
+  - @px-ui/core@4.4.2
+
+## 11.0.1
+
+### Patch Changes
+
+- Updated dependencies [e682c07]
+  - @px-ui/core@4.4.1
+
+## 11.0.0
+
+### Minor Changes
+
+- d09d505: add onRemove Callback for file upload Clear files component, update the currency dropdown to be able to search currency via country name
+
+### Patch Changes
+
+- Updated dependencies [d09d505]
+  - @px-ui/core@4.4.0
+
+## 10.0.1
+
+### Patch Changes
+
+- Updated dependencies [465106a]
+  - @px-ui/core@4.3.1
+
+## 10.0.0
+
+### Minor Changes
+
+- bd0e49c: added onFileRemove and onFileAdd to the useFileUpload hook
+
+### Patch Changes
+
+- Updated dependencies [bd0e49c]
+  - @px-ui/core@4.3.0
+
 ## 9.0.0
 
 ### Minor Changes